Delete README.md
Email marketing is broadcasting commercial messages to a group of people using email. Currently, email marketing is consistently delivering relatively high return on investment (ROI) in a marketing field. However, collecting subscriber information and sending the email only to interested consumers is a major research issue in email marketing.  E-mail marketing is compared with other forms of direct and Internet marketing, identifying its key advantages. Design a model that will be able to predict whether a customer will respond to the marketing campaign based on his/her information. In other words, predict the ‘responded’ target variable described above based on all the input variables provided.

Email marketing has been proved a great medium in order to get  either a  lead or profit. Out  of 100 emails  if we are possible to convert it into to one sale then also it would be a profitable business. In today time when email marketing is touching its peak users are still facing some huge problems like spam, bounce, SMTP suspension, and etc. So where the actual problem is either ESP services are not good or users do  not  to  proper  use  of  ESP  for  sending  the  email campaign. Nowadays most of the ESP are using a complex email filter which can’t we assured for sending our email in the user’s box.